refactoring the code for address list and address detail module.
