+++ console/console-init/ui/src/Tests/EmptyAddress.test.tsx
@@ -5,7 +5,7 @@
 
 import * as React from "react";
 import { render } from "@testing-library/react";
-import { EmptyAddress } from "components/AddressSpace/Address/EmptyAddress";
+import { EmptyAddress } from "modules/address/components/EmptyAddress";
 
 describe("Empty Page with Empty state for Addresses", () => {
   test("it renders empty address state instead of list of addresses", () => {
