+++ console/console-init/ui/src/components/AddressSpace/Connection/ConnectionList.tsx
@@ -17,7 +17,7 @@ import { Link } from "react-router-dom";
 import { ConnectionProtocolFormat } from "components/common/ConnectionListFormatter";
 import useWindowDimensions from "components/common/WindowDimension";
 import { FormatDistance } from "use-patternfly";
-import { StyleForTable } from "modules/address-space/components/AddressSpaceList";
+import { StyleForTable } from "modules/address-space/components/AddressSpaceList/AddressSpaceList";
 import { css } from "@patternfly/react-styles";
 
 interface IConnectionListProps {
