+++ console/console-init/ui/src/components/RootModal.tsx
@@ -5,10 +5,12 @@ import {
   EditAddressSpace,
   CreateAddressSpace
 } from "modules/address-space/dialogs";
+import { EditAddress } from "modules/address/containers";
 
 const MODAL_COMPONENTS: any = {
   [MODAL_TYPES.CREATE_ADDRESS_SPACE]: CreateAddressSpace,
   [MODAL_TYPES.EDIT_ADDRESS_SPACE]: EditAddressSpace,
+  [MODAL_TYPES.EDIT_ADDRESS]: EditAddress,
   [MODAL_TYPES.DELETE_ADDRESS_SPACE]: DialogPrompt,
   [MODAL_TYPES.DELETE_ADDRESS]: DialogPrompt,
   [MODAL_TYPES.PURGE_ADDRESS]: DialogPrompt
