+++ console/console-init/ui/src/components/common/index.ts
@@ -4,7 +4,6 @@
  */
 
 export * from "./About";
-export * from "./AddressFormatter";
 export * from "./AddressSpaceListFormatter";
 export * from "./ConnectionListFormatter";
 export * from "./DialogPrompt";
@@ -12,10 +11,8 @@ export * from "./EmptyLinks";
 export * from "./Error";
 export * from "./ErrorBoundary";
 export * from "./FilterDropdown";
-export * from "./Messages";
 export * from "./NotFound";
 export * from "./ServerMessageAlert";
 export * from "./SortForMobileView";
-export * from "./TypePlan";
 export * from "./WindowDimension";
 export * from "./NetworkStatusAlert";
