+++ console/console-init/ui/src/context-state-reducer/actions.ts
@@ -15,6 +15,7 @@ const HIDE_MODAL = "HIDE_MODAL";
  */
 const DELETE_ADDRESS_SPACE = "DELETE_ADDRESS_SPACE";
 const EDIT_ADDRESS_SPACE = "EDIT_ADDRESS_SPACE";
+const EDIT_ADDRESS = "EDIT_ADDRESS";
 const CREATE_ADDRESS_SPACE = "CREATE_ADDRESS_SPACE";
 
 /**
@@ -35,6 +36,7 @@ export const types = {
 export const MODAL_TYPES = {
   DELETE_ADDRESS_SPACE,
   EDIT_ADDRESS_SPACE,
+  EDIT_ADDRESS,
   CREATE_ADDRESS_SPACE,
   DELETE_ADDRESS,
   PURGE_ADDRESS
