+++ console/console-init/ui/src/modules/address-detail/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./AddressDetailHeader";
+export * from "./AddressLinks";
+export * from "./EmptyAddressLinks";
