+++ console/console-init/ui/src/modules/address-detail/containers/index.ts
@@ -3,4 +3,6 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./utils";
+export * from "./AddressLinksContainer";
+export * from "./AddressLinksToolbar";
+export * from "./AddressLinksListPage";
