+++ console/console-init/ui/src/modules/address-detail/index.ts
