+++ console/console-init/ui/src/modules/address-detail/utils/index.ts
@@ -1,0 +1,1 @@
+export {};
