+++ console/console-init/ui/src/modules/address-space/components/AddressSpaceList/AddressSpaceList.tsx
@@ -26,6 +26,7 @@ import { IAddressSpacesResponse } from "types/ResponseTypes";
 import { FetchPolicy, POLL_INTERVAL } from "constants/constants";
 import { IObjectMeta_v1_Input } from "pages/AddressSpaceDetail/AddressSpaceDetailPage";
 import { useMutationQuery } from "hooks";
+import { compareTwoAddress } from "modules/address/utils";
 import { useStoreContext, types, MODAL_TYPES } from "context-state-reducer";
 import { compareObject } from "utils";
 import {
@@ -77,6 +78,7 @@ interface AddressSpaceListProps {
   ) => void;
   selectedAddressSpaces: Array<IAddressSpace>;
 }
+
 export const AddressSpaceList: React.FunctionComponent<AddressSpaceListProps> = ({
   page,
   perPage,
