+++ console/console-init/ui/src/modules/address/components/AddressErrorMessage.tsx
@@ -1,0 +1,21 @@
+import * as React from "react";
+
+interface IAddressErrorMessageProps {
+  messages: Array<string>;
+}
+
+export const AddressErrorMessage: React.FunctionComponent<IAddressErrorMessageProps> = ({
+  messages
+}) => {
+  return (
+    <>
+      {messages &&
+        messages.map(message => (
+          <>
+            {message}
+            <br />
+          </>
+        ))}
+    </>
+  );
+};
