+++ console/console-init/ui/src/modules/address/components/index.ts
@@ -1,1 +1,9 @@
-export {};
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+export * from "./EmptyAddress";
+export * from "./AddressList";
+export * from "./Messages";
+export * from "./TypePlan";
