+++ console/console-init/ui/src/modules/address/containers/index.ts
@@ -1,1 +1,12 @@
-export {};
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+export * from "./AddressListContainer";
+export * from "./AddressListFilter";
+export * from "./AddressListToolbar";
+export * from "./EditAddressPage";
+export * from "./CreateAddress/CreateAddressDefinition";
+export * from "./CreateAddress/CreateAddressPage";
+export * from "./CreateAddress/PreviewAddress";
